my project
The main goal of this project is to design and implement a traffic
security system based on Raspberry Pi 3(latest version of raspberry pi
device) which can be deployed on the traffic red lights and highways.
Using this system, we can detect if there’s water logging on the road,
if there’s fire on the road, condition of the road, and anyone who is
breaking the traffic red light.

The main advantage of traffic security system is it is better than any
other security systems. This system consist of a Water Sensor(which
sense if there’s water logging on the road), an IR sensor(activated when
light is red, and checks if anyone is breaking it), a Smoke Sensor(to
detect
If there’s fire on the road) and traffic red light(green and red LED’s),
a Web Camera to instantly click photos of any situation. .Whenever any
of these sensor is activated we receive a message on our device (cell
phone) and the PUTTY software we are using is used to activate our
raspberry pi device and system. We also use a drop box which stores all
the pictures that our webcam is clicking. Therefore with the help of
this message and picture system we can instantly notify the specific
authority so they can respond as soon as possible.